IN-250 refactor: change file structure
