+++ src/features/createPost/model/stores/index.ts
@@ -1,5 +1,2 @@
 export * from './addPostStore'
-export * from './cropStore'
-export * from './filterStore'
-export * from './imageCollection'
-export * from './photoStore'
+export * from './imageStore'
